requirements.txt
requests==2.26.0
strconv==0.4.2
python-dateutil==2.8.2
jsons==1.6.0
flask-restful==0.3.9
flask-cors==3.0.10
werkzeug==2.0.2